Relocate DNS classes to zeroconf.dns
